Django_PostgreSQL/api/requirements.txt
Django>=4.2
django-model-utils>=4.3
djangorestframework>=3.14
psycopg2>=2.9
gunicorn>=20.1